dom/canvas/test/test_windingRuleUndefined.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=861938
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 861938</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=861938">Mozilla Bug 861938</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
  <script type="application/javascript">

  /** Test for Bug 861938 **/

  try {
    var canvas = document.createElement("canvas");
    canvas.getContext('2d').fill(undefined);

    ok(true, "Should be able to pass undefined to fill()");
  } catch(e) {
    ok(false, "Should be able to pass undefined to fill()");
  }
  </script>
</body>
</html>
